My Data Structure and Procedural Programming README.md
PROBLEM 1:
Distinct Elements Sum Algorithm:
In Problem 1, this JavaScript algorithm is designed to find the sum of all distinct elements from two sets of elements.
It calculates the sum of elements that are present in either of the given sets.
The algorithm utilizes arrays for representing the sets and follows a straightforward approach to achieve the desired result.

PROBLEM 2:
In problem 2, we Write an algorithm that determines, for n pairs of given vectors, whether two vectors of real numbers are orthogonal. This is done by calling the dot_product function defined in the previous question. The dot product of two orthogonal vectors is zero.